+++ src/app/core/services/api/fake-backend/connector-fake-impl/data/test-assets.ts
@@ -73,6 +73,21 @@ export namespace TestAssets {
 This is a short description text that should be fully rendered without being **collapsed**. No *show more* button should be visible.
 `;
 
+  export const onRequestAsset: UiAsset = {
+    dataSourceAvailability: 'ON_REQUEST',
+    assetId: 'comfee-or',
+    title: 'Comfee OR',
+    description: '',
+    descriptionShortText: '',
+    connectorEndpoint: 'https://my-other-connector/api/dsp',
+    participantId: 'MDSL1234XX.C1234XX',
+    creatorOrganizationName: 'my-org',
+    temporalCoverageFrom: new Date('2024-01-01'),
+    onRequestContactEmail: 'mail@mail.mail',
+    onRequestContactEmailSubject: "Request for asset 'comfee-or'",
+    isOwnConnector: false,
+  };
+
   export const boring: UiAsset = {
     dataSourceAvailability: 'LIVE',
     assetId: 'data-sample-ckd-skd-demands-2023-Jan',