+++ src/app/core/services/api/fake-backend/connector-fake-impl/ee-fake-service.ts
@@ -2,10 +2,10 @@ import {ConnectorLimits} from '@sovity.de/edc-client';
 import {contractAgreementPage} from './contract-agreement-fake-service';
 
 export const connectorLimits = (): ConnectorLimits => ({
-  numActiveConsumingContractAgreements: contractAgreementPage(
-    {},
-  ).contractAgreements.filter(
-    (it) => it.direction === 'CONSUMING' && it.terminationStatus === 'ONGOING',
-  ).length,
+  numActiveConsumingContractAgreements:
+    contractAgreementPage().contractAgreements.filter(
+      (it) =>
+        it.direction === 'CONSUMING' && it.terminationStatus === 'ONGOING',
+    ).length,
   maxActiveConsumingContractAgreements: 1,
 });