+++ CHANGELOG.md
@@ -20,6 +20,7 @@ the detailed section referring to by linking pull requests or issues.
 
 - Both providers and consumers can now terminate contracts.
 - Contracts can be filtered by their termination status.
+- Adjusted data offer card/detail dialog UI to differentiate live and on request assets
 
 #### Patch